Refactoring : 방 참여 페이지 리팩토링
